requirements.txt
google-api-core==2.8.2
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.56.4